+++ src/CleanRegex/Match/Details/Group/NotMatchedGroup.php
@@ -1,11 +1,9 @@
 <?php
 namespace TRegx\CleanRegex\Match\Details\Group;
 
-use Throwable;
 use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Internal\GroupEmptyOptional;
 use TRegx\CleanRegex\Internal\Match\Details\Group\GroupDetails;
-use TRegx\CleanRegex\Internal\Message\GroupMessage;
 use TRegx\CleanRegex\Internal\Subject;
 use TRegx\CleanRegex\Match\Details\NotMatched;
 use TRegx\CleanRegex\Match\Optional;
@@ -18,15 +16,12 @@ class NotMatchedGroup implements Group
     private $details;
     /** @var NotMatched */
     private $notMatched;
-    /** @var GroupMessage */
-    private $groupMessage;
 
     public function __construct(Subject $subject, GroupDetails $details, NotMatched $notMatched)
     {
         $this->subject = $subject;
         $this->details = $details;
         $this->notMatched = $notMatched;
-        $this->groupMessage = new GroupMessage($this->details->group());
     }
 
     public function text(): string
@@ -128,14 +123,6 @@ class NotMatchedGroup implements Group
     public function orReturn($substitute)
     {
         return $substitute;
-    }
-
-    public function orThrow(Throwable $throwable = null): void
-    {
-        if ($throwable === null) {
-            throw new GroupNotMatchedException($this->groupMessage->getMessage());
-        }
-        throw $throwable;
     }
 
     public function orElse(callable $substituteProducer)