+++ test/Fakes/CleanRegex/Match/Details/Group/ThrowGroup.php
@@ -2,7 +2,6 @@
 namespace Test\Fakes\CleanRegex\Match\Details\Group;
 
 use Test\Utils\Fails;
-use Throwable;
 use TRegx\CleanRegex\Match\Details\Group\Group;
 use TRegx\CleanRegex\Match\Optional;
 
@@ -91,11 +90,6 @@ class ThrowGroup implements Group
     }
 
     public function index(): int
-    {
-        throw $this->fail();
-    }
-
-    public function orThrow(Throwable $throwable = null)
     {
         throw $this->fail();
     }